Express & Mongoose refactoring for jest setup
