+++ src/server/app.js
@@ -1,0 +1,14 @@
+import "core-js/stable";
+import "regenerator-runtime/runtime";
+
+import express from "express";
+import bodyParser from "body-parser";
+
+import { route } from "./controllers";
+
+const app = express();
+
+app.use(bodyParser.json());
+route(app);
+
+export default app;
